% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innerApply.R
\name{innerApply}
\alias{innerApply}
\title{Apply a function to an atomic vector using its two elements each time}
\usage{
innerApply(input, .f0, distance = 1)
}
\arguments{
\item{input}{An atomic vector.}

\item{.f0}{A function. Anonymous functions are supported. Formula or vectors are not.}

\item{distance}{The distance between each element of the vector. The default value is 1.}
}
\value{
An atomic vector of length equals length(input)-distance.
}
\description{
#' The innerApply function applies a function to an atomic vector using two of its elements in sequence as inputs. It is different from purrr::accumulate since it does not use any of its output as an input.
}
\details{
If the function used as .f0 returns a vector of length higher than 1, the function will throw an error.
}
\examples{


inputVec <- 1:30

outputVec <- innerApply(inputVec, `+`)

innerApply(letters, paste0)

innerApply(c(TRUE,FALSE,TRUE,FALSE), `!=`)
}
